refactor: consolidate service factories into api/deps.py (#681)
Move all get_*_service() factory functions from individual service modules
into a single wikimind/api/services.py module. Routes and tests now import
factories from one location. Circular imports between api/services and
jobs/worker/engine/linter are resolved with lazy imports.


[skip-doc-check]
