+++ src/wikimind/api/routes/compilation_schemas.py
@@ -5,16 +5,14 @@ from fastapi import APIRouter, Depends
 from sqlmodel.ext.asyncio.session import AsyncSession
 
 from wikimind.api.deps import get_current_user_id
+from wikimind.api.services import get_compilation_schema_service
 from wikimind.database import get_session
 from wikimind.models import (
     CompilationSchemaResponse,
     CreateCompilationSchemaRequest,
     UpdateCompilationSchemaRequest,
 )
-from wikimind.services.compilation_schema import (
-    CompilationSchemaService,
-    get_compilation_schema_service,
-)
+from wikimind.services.compilation_schema import CompilationSchemaService
 
 log = structlog.get_logger()
