+++ src/wikimind/api/routes/ingest.py
@@ -9,6 +9,7 @@ from sqlmodel import select
 from sqlmodel.ext.asyncio.session import AsyncSession
 
 from wikimind.api.deps import get_current_user_id
+from wikimind.api.services import get_ingest_service
 from wikimind.database import get_session
 from wikimind.ingest.adapters.pdf import PDFAdapter
 from wikimind.models import (
@@ -18,7 +19,7 @@ from wikimind.models import (
     SourceContentResponse,
     SourceImage,
 )
-from wikimind.services.ingest import IngestService, get_ingest_service
+from wikimind.services.ingest import IngestService
 from wikimind.storage import find_original_sibling, get_raw_storage
 
 # Strict pattern: only alphanumeric, hyphens, underscores, dots (no path separators)
