+++ src/wikimind/api/routes/lint.py
@@ -7,6 +7,7 @@ from typing import TYPE_CHECKING
 from fastapi import APIRouter, Depends, Query
 
 from wikimind.api.deps import get_current_user_id
+from wikimind.api.services import get_linter_service
 from wikimind.database import get_session
 from wikimind.models import (
     DismissFindingResponse,
@@ -15,10 +16,11 @@ from wikimind.models import (
     LintReportDetail,
     LintRunResponse,
 )
-from wikimind.services.linter import LinterService, get_linter_service
 
 if TYPE_CHECKING:
     from sqlmodel.ext.asyncio.session import AsyncSession
+
+    from wikimind.services.linter import LinterService
 
 router = APIRouter()
