+++ src/wikimind/api/routes/tags.py
@@ -5,13 +5,14 @@ from fastapi import APIRouter, Depends
 from sqlmodel.ext.asyncio.session import AsyncSession
 
 from wikimind.api.deps import get_current_user_id
+from wikimind.api.services import get_tag_service, get_wiki_service
 from wikimind.database import get_session
 from wikimind.models import (
     CreateTagRequest,
     TagResponse,
 )
-from wikimind.services.tags import TagService, get_tag_service
-from wikimind.services.wiki import WikiService, get_wiki_service
+from wikimind.services.tags import TagService
+from wikimind.services.wiki import WikiService
 
 log = structlog.get_logger()
