+++ tests/unit/test_search.py
@@ -8,6 +8,7 @@ from sqlmodel import SQLModel
 from sqlmodel.ext.asyncio.session import AsyncSession
 
 from tests.conftest import TEST_USER_ID
+from wikimind.api.services import get_search_service
 from wikimind.models import Article, SearchResponse, SearchResult, User
 from wikimind.services.search import (
     SearchService,
@@ -15,7 +16,6 @@ from wikimind.services.search import (
     _sanitize_fts5_query,
     _sanitize_postgres_query,
     create_fts_table,
-    get_search_service,
     index_article,
     rebuild_fts_index,
     remove_article,
