+++ tests/unit/test_tag_service.py
@@ -7,9 +7,10 @@ from typing import TYPE_CHECKING
 import pytest
 
 from tests.conftest import TEST_USER_ID
+from wikimind.api.services import get_tag_service
 from wikimind.errors import NotFoundError
 from wikimind.models import Article, PageType
-from wikimind.services.tags import TagService, get_tag_service
+from wikimind.services.tags import TagService
 
 if TYPE_CHECKING:
     from sqlmodel.ext.asyncio.session import AsyncSession
